ANDKR-1495 | 취소내역 화면 개발 - 취소 내역 상세 연결 제외 작업 완료

+++ dailyHOTEL/app/src/main/java/com/daily/dailyhotel/entity/BookingCancel.java
@@ -3,7 +3,7 @@ package com.daily.dailyhotel.entity;
 public class BookingCancel
 {
     public String aggregationId;
-    public int orderSeq;
+    public long orderSeq;
     public int reservationIdx;
 //    public String checkinDate;
 //    public String checkoutDate;
